+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/environment-check
 
+## 0.16.37
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.19.1
+-   @sap-ux/axios-extension@1.11.3
+
+## 0.16.36
+
+### Patch Changes
+
+-   Updated dependencies [a4f00f7c]
+    -   @sap-ux/project-access@1.19.0
+    -   @sap-ux/axios-extension@1.11.3
+
 ## 0.16.35
 
 ### Patch Changes