+++ packages/telemetry/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/telemetry
 
+## 0.4.6
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.19.1
+
+## 0.4.5
+
+### Patch Changes
+
+-   Updated dependencies [a4f00f7c]
+    -   @sap-ux/project-access@1.19.0
+
 ## 0.4.4
 
 ### Patch Changes